refactor and fix boost.python interface
