+++ api/boostpython/boostpython.cbp
@@ -91,6 +91,10 @@
 			<Option target="api_Debug" />
 			<Option target="api_Release" />
 		</Unit>
+		<Unit filename="api_utctime.cpp">
+			<Option target="api_Debug" />
+			<Option target="api_Release" />
+		</Unit>
 		<Unit filename="pt_gs_k.cpp">
 			<Option target="pt_gs_k_Debug" />
 			<Option target="pt_gs_k_Release" />
